client: split Coord in two struct Map / World
- update coordinate code using smallest integer for the need
- World use i8 because we only have 2 maps for now and we will never exceed 4 maps
- Map use i16 because i8::MAX is not enough
(e.g. TILEMAP_WIDTH x TILEMAP_HEIGHT = 16 * 12 = 192)
- precompute TILEMAP_LINEAR_SIZE in constants

+++ client/src/entities/view.rs
@@ -59,8 +59,8 @@ impl EntityView {
                             let cell_color = [0.0, (1.0 - (1.0 / index as f32)), 1.0, 0.4];
                             Rectangle::new(cell_color).draw(
                                 [
-                                    self.margin.width + (cell.x * 64 as i32) as f64,
-                                    self.margin.height + (cell.y * 64 as i32) as f64,
+                                    self.margin.width + (cell.x as f64 * 64.0),
+                                    self.margin.height + (cell.y as f64 * 64.0),
                                     64.0,
                                     64.0,
                                 ],
